refactor(arxiv): optimize metadata handling in ArxivDocumentReader
- Extract createMetadata method for unified metadata creation
- Add addIfNotNull and addIfNotEmpty helper methods
- Simplify metadata addition logic for better maintainability
- Standardize null value and empty collection checks
- Improve code structure for better reusability
